+++ binless/R/binless.R
@@ -28,8 +28,11 @@
 #' \code{\link{plot_binless_matrix}}
 #'
 #' @useDynLib binless, .registration = TRUE
+#' @import Rcpp
+#' @import RcppEigen
 #' @import methods
 #' @import data.table
+#' @import parallel
 #' @import doParallel
 #' @import foreach
 #' @import matrixStats